pleiiin de trucs

+++ src/controllers/typeEvent-supprimer.php
@@ -1,7 +1,8 @@
 <?php
+
 require_once(MODELS_INC."TypeEvenementDAO.class.php");
 
-if ($_GET['id']) {
+if ($_GET['id'] && $_SESSION['user_auth']['write']) {
 	$typeEvent = new TypeEvenement($_GET['id'], NULL);
 	TypeEvenementDAO::delete($typeEvent);
 	header('Location: '.SELF.'typesEvent');
